refactor: refactor home folder structure
