+++ src/app/(protected)/(tabs)/home/index.tsx
@@ -1,1 +1,1 @@
-export { Home as default } from '~/screens/home-screen'
+export { Home as default } from '~/modules/home/screens/home-screen'
