+++ src/modules/home/parts/quick-action-item/index.static.ts
